+++ src/utils/MessageHelper/MessageBuilder.ts
@@ -116,11 +116,66 @@ export default class MessageBuilder {
   }
 
   static async activeMatchMessage({ blue, orange }: ActiveMatchCreated): Promise<MessageOptions> {
-    const embed = await EmbedBuilder.activeMatchEmbed({ blue, orange });
+    //const embed = await EmbedBuilder.activeMatchEmbed({ blue, orange });
+    const blueTeam: Array<string> = blue.players.map((player) => "<@" + player.id + ">");
+    const orangeTeam: Array<string> = orange.players.map((player) => "<@" + player.id + ">");
+    const activeMatchEmbed = new MessageEmbed({
+      color: ColorCodes.DarkRed,
+      fields: [
+        { name: "🔷 Blue Team 🔷", value: blueTeam.join("\n") },
+        { name: "🔶 Orange Team 🔶", value: orangeTeam.join("\n") },
+      ],
+      thumbnail: { url: this.normIconURL },
+      title: "Teams are set!",
+    });
+
+    let probability;
+    let winner;
+    if (blue.winProbability > orange.winProbability) {
+      probability = blue.winProbability;
+      winner = "Blue Team is";
+    } else if (blue.winProbability < orange.winProbability) {
+      probability = orange.winProbability;
+      winner = "Orange Team is";
+    } else {
+      probability = "50";
+      winner = "Both teams are";
+    }
+
+    const event = await EventRepository.getCurrentEvent();
+    const blueMMR = blue.mmrStake * event.mmrMult;
+    const orangeMMR = orange.mmrStake * event.mmrMult;
+
+    activeMatchEmbed.addFields({
+      name: "MMR Stake & Probability Rating:\n",
+      value:
+        "🔷 Blue Team: \u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0\u00A0**(+" +
+        blueMMR.toString() +
+        ")**\u00A0\u00A0**(-" +
+        orange.mmrStake.toString() +
+        ")** 🔷\n🔶 Orange Team:\u00A0\u00A0**(+" +
+        orangeMMR.toString() +
+        ")**\u00A0\u00A0**(-" +
+        blue.mmrStake.toString() +
+        ")** 🔶\n" +
+        winner +
+        " predicted to have a **" +
+        probability +
+        "%** chance of winning.",
+    });
+
+    if (event.mmrMult > 1) {
+      activeMatchEmbed.addFields({
+        name: "X" + event.mmrMult.toString() + " MMR Event!",
+        value: "Winnings are multiplied by **" + event.mmrMult.toString() + "** for this match!",
+      });
+    }
+
+    activeMatchEmbed.addFields({ name: "Reporting", value: "Use the buttons to report which team won the match." });
 
     return {
       components: [ButtonBuilder.activeMatchButtons()],
-      embeds: [embed],
+      embeds: [activeMatchEmbed],
     };
   }
